Add is-woo-passwordless class at layout component
